+++ include/Components/Player.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <entityx/Entity.h>
+
+namespace Spectacle
+{
+	struct Player : public entityx::Component< Player >
+	{
+		float moveSpeed;
+
+		Player( float speed = 1.0f );
+	};
+}
