+++ include/Spectacle.h
@@ -6,4 +6,6 @@
 #include <Components/FollowTargetLinear.h>
 #include <Components/Mesh.h>
 #include <Components/Light.h>
+#include <Components/Player.h>
 #include <Systems/FollowTargetLinearSystem.h>
+#include <Systems/PlayerMovementSystem.h>
