+++ source/Components/Player.cpp
@@ -1,0 +1,9 @@
+#include "Components/Player.h"
+
+namespace Spectacle
+{
+	Player::Player( float speed )
+		: moveSpeed( speed )
+	{
+	}
+}
